requirements.txt
Mastodon.py==1.2.1
markovify==0.7.1
